# realworld/api/shop/v1/shop_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.2
// - protoc             v5.29.3
// source: shop/v1/shop.proto

package v1

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationShopServiceBuy = "/shop.v1.ShopService/Buy"
const OperationShopServiceListItems = "/shop.v1.ShopService/ListItems"

type ShopServiceHTTPServer interface {
	// Buy POST /api/v1/shop/buy 购买
	Buy(context.Context, *BuyRequest) (*BuyReply, error)
	// ListItems GET /api/v1/shop/items 商品列表，query: category 可选
	ListItems(context.Context, *ListItemsRequest) (*ListItemsReply, error)
}

func RegisterShopServiceHTTPServer(s *http.Server, srv ShopServiceHTTPServer) {
	r := s.Route("/")
	r.GET("/api/v1/shop/items", _ShopService_ListItems0_HTTP_Handler(srv))
	r.POST("/api/v1/shop/buy", _ShopService_Buy0_HTTP_Handler(srv))
}

func _ShopService_ListItems0_HTTP_Handler(srv ShopServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in ListItemsRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationShopServiceListItems)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.ListItems(ctx, req.(*ListItemsRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*ListItemsReply)
		return ctx.Result(200, reply)
	}
}

func _ShopService_Buy0_HTTP_Handler(srv ShopServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in BuyRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationShopServiceBuy)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Buy(ctx, req.(*BuyRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*BuyReply)
		return ctx.Result(200, reply)
	}
}

type ShopServiceHTTPClient interface {
	// Buy POST /api/v1/shop/buy 购买
	Buy(ctx context.Context, req *BuyRequest, opts ...http.CallOption) (rsp *BuyReply, err error)
	// ListItems GET /api/v1/shop/items 商品列表，query: category 可选
	ListItems(ctx context.Context, req *ListItemsRequest, opts ...http.CallOption) (rsp *ListItemsReply, err error)
}

type ShopServiceHTTPClientImpl struct {
	cc *http.Client
}

func NewShopServiceHTTPClient(client *http.Client) ShopServiceHTTPClient {
	return &ShopServiceHTTPClientImpl{client}
}

// Buy POST /api/v1/shop/buy 购买
func (c *ShopServiceHTTPClientImpl) Buy(ctx context.Context, in *BuyRequest, opts ...http.CallOption) (*BuyReply, error) {
	var out BuyReply
	pattern := "/api/v1/shop/buy"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationShopServiceBuy))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// ListItems GET /api/v1/shop/items 商品列表，query: category 可选
func (c *ShopServiceHTTPClientImpl) ListItems(ctx context.Context, in *ListItemsRequest, opts ...http.CallOption) (*ListItemsReply, error) {
	var out ListItemsReply
	pattern := "/api/v1/shop/items"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationShopServiceListItems))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}