# realworld/api/wechat/v1/wechat_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.2
// - protoc             v5.29.3
// source: wechat/v1/wechat.proto

package v1

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
	emptypb "google.golang.org/protobuf/types/known/emptypb"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationWechatServiceGetLoginQrcode = "/wechat.v1.WechatService/GetLoginQrcode"
const OperationWechatServiceGetLoginStatus = "/wechat.v1.WechatService/GetLoginStatus"
const OperationWechatServiceSendGroupImage = "/wechat.v1.WechatService/SendGroupImage"
const OperationWechatServiceSendGroupMentionText = "/wechat.v1.WechatService/SendGroupMentionText"
const OperationWechatServiceSendGroupText = "/wechat.v1.WechatService/SendGroupText"

type WechatServiceHTTPServer interface {
	GetLoginQrcode(context.Context, *emptypb.Empty) (*GetLoginQrcodeReply, error)
	GetLoginStatus(context.Context, *emptypb.Empty) (*GetLoginStatusReply, error)
	SendGroupImage(context.Context, *SendGroupImageRequest) (*SendGroupImageReply, error)
	SendGroupMentionText(context.Context, *SendGroupMentionTextRequest) (*SendGroupMentionTextReply, error)
	SendGroupText(context.Context, *SendGroupTextRequest) (*SendGroupTextReply, error)
}

func RegisterWechatServiceHTTPServer(s *http.Server, srv WechatServiceHTTPServer) {
	r := s.Route("/")
	r.POST("/api/wechat/v1/group/text", _WechatService_SendGroupText0_HTTP_Handler(srv))
	r.POST("/api/wechat/v1/group/image", _WechatService_SendGroupImage0_HTTP_Handler(srv))
	r.POST("/api/wechat/v1/group/mention", _WechatService_SendGroupMentionText0_HTTP_Handler(srv))
	r.GET("/api/wechat/v1/login/qrcode", _WechatService_GetLoginQrcode0_HTTP_Handler(srv))
	r.GET("/api/wechat/v1/login/status", _WechatService_GetLoginStatus0_HTTP_Handler(srv))
}

func _WechatService_SendGroupText0_HTTP_Handler(srv WechatServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in SendGroupTextRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationWechatServiceSendGroupText)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.SendGroupText(ctx, req.(*SendGroupTextRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*SendGroupTextReply)
		return ctx.Result(200, reply)
	}
}

func _WechatService_SendGroupImage0_HTTP_Handler(srv WechatServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in SendGroupImageRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationWechatServiceSendGroupImage)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.SendGroupImage(ctx, req.(*SendGroupImageRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*SendGroupImageReply)
		return ctx.Result(200, reply)
	}
}

func _WechatService_SendGroupMentionText0_HTTP_Handler(srv WechatServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in SendGroupMentionTextRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationWechatServiceSendGroupMentionText)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.SendGroupMentionText(ctx, req.(*SendGroupMentionTextRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*SendGroupMentionTextReply)
		return ctx.Result(200, reply)
	}
}

func _WechatService_GetLoginQrcode0_HTTP_Handler(srv WechatServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in emptypb.Empty
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationWechatServiceGetLoginQrcode)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetLoginQrcode(ctx, req.(*emptypb.Empty))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*GetLoginQrcodeReply)
		return ctx.Result(200, reply)
	}
}

func _WechatService_GetLoginStatus0_HTTP_Handler(srv WechatServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in emptypb.Empty
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationWechatServiceGetLoginStatus)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetLoginStatus(ctx, req.(*emptypb.Empty))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*GetLoginStatusReply)
		return ctx.Result(200, reply)
	}
}

type WechatServiceHTTPClient interface {
	GetLoginQrcode(ctx context.Context, req *emptypb.Empty, opts ...http.CallOption) (rsp *GetLoginQrcodeReply, err error)
	GetLoginStatus(ctx context.Context, req *emptypb.Empty, opts ...http.CallOption) (rsp *GetLoginStatusReply, err error)
	SendGroupImage(ctx context.Context, req *SendGroupImageRequest, opts ...http.CallOption) (rsp *SendGroupImageReply, err error)
	SendGroupMentionText(ctx context.Context, req *SendGroupMentionTextRequest, opts ...http.CallOption) (rsp *SendGroupMentionTextReply, err error)
	SendGroupText(ctx context.Context, req *SendGroupTextRequest, opts ...http.CallOption) (rsp *SendGroupTextReply, err error)
}

type WechatServiceHTTPClientImpl struct {
	cc *http.Client
}

func NewWechatServiceHTTPClient(client *http.Client) WechatServiceHTTPClient {
	return &WechatServiceHTTPClientImpl{client}
}

func (c *WechatServiceHTTPClientImpl) GetLoginQrcode(ctx context.Context, in *emptypb.Empty, opts ...http.CallOption) (*GetLoginQrcodeReply, error) {
	var out GetLoginQrcodeReply
	pattern := "/api/wechat/v1/login/qrcode"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationWechatServiceGetLoginQrcode))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *WechatServiceHTTPClientImpl) GetLoginStatus(ctx context.Context, in *emptypb.Empty, opts ...http.CallOption) (*GetLoginStatusReply, error) {
	var out GetLoginStatusReply
	pattern := "/api/wechat/v1/login/status"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationWechatServiceGetLoginStatus))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *WechatServiceHTTPClientImpl) SendGroupImage(ctx context.Context, in *SendGroupImageRequest, opts ...http.CallOption) (*SendGroupImageReply, error) {
	var out SendGroupImageReply
	pattern := "/api/wechat/v1/group/image"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationWechatServiceSendGroupImage))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *WechatServiceHTTPClientImpl) SendGroupMentionText(ctx context.Context, in *SendGroupMentionTextRequest, opts ...http.CallOption) (*SendGroupMentionTextReply, error) {
	var out SendGroupMentionTextReply
	pattern := "/api/wechat/v1/group/mention"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationWechatServiceSendGroupMentionText))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *WechatServiceHTTPClientImpl) SendGroupText(ctx context.Context, in *SendGroupTextRequest, opts ...http.CallOption) (*SendGroupTextReply, error) {
	var out SendGroupTextReply
	pattern := "/api/wechat/v1/group/text"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationWechatServiceSendGroupText))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}