# realworld/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /api/v1/shop/buy:
        post:
            tags:
                - ShopService
            description: POST /api/v1/shop/buy 购买
            operationId: ShopService_Buy
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/shop.v1.BuyRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/shop.v1.BuyReply'
    /api/v1/shop/items:
        get:
            tags:
                - ShopService
            description: 'GET /api/v1/shop/items 商品列表，query: category 可选'
            operationId: ShopService_ListItems
            parameters:
                - name: category
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/shop.v1.ListItemsReply'
...
    /api/wechat/v1/group/image:
        post:
            tags:
                - WechatService
            operationId: WechatService_SendGroupImage
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/wechat.v1.SendGroupImageRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/wechat.v1.SendGroupImageReply'
    /api/wechat/v1/group/mention:
        post:
            tags:
                - WechatService
            operationId: WechatService_SendGroupMentionText
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/wechat.v1.SendGroupMentionTextRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/wechat.v1.SendGroupMentionTextReply'
    /api/wechat/v1/group/text:
        post:
            tags:
                - WechatService
            operationId: WechatService_SendGroupText
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/wechat.v1.SendGroupTextRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/wechat.v1.SendGroupTextReply'
...
components:
    schemas:
...
        shop.v1.BuyReply:
            type: object
            properties:
                success:
                    type: boolean
                message:
                    type: string
        shop.v1.BuyRequest:
            type: object
            properties:
                item_id:
                    type: string
                quantity:
                    type: integer
                    format: int32
            description: Buy 购买：扣用户金币，将商品写入背包
        shop.v1.ListItemsReply:
            type: object
            properties:
                items:
                    type: array
                    items:
                        $ref: '#/components/schemas/shop.v1.ShopItem'
        shop.v1.ShopItem:
            type: object
            properties:
                id:
                    type: string
                name:
                    type: string
                category:
                    type: string
                price_coins:
                    type: integer
                    format: int32
                image_url:
                    type: string
                description:
                    type: string
            description: ShopItem 商品信息，用于列表与购买结果
...
        wechat.v1.SendGroupImageReply:
            type: object
            properties:
                success:
                    type: boolean
                messageId:
                    type: string
        wechat.v1.SendGroupImageRequest:
            type: object
            properties:
                roomId:
                    type: string
                roomTopic:
                    type: string
                imageUrl:
                    type: string
                imageBytes:
                    type: string
                    format: bytes
                filename:
                    type: string
        wechat.v1.SendGroupMentionTextReply:
            type: object
            properties:
                success:
                    type: boolean
                messageId:
                    type: string
        wechat.v1.SendGroupMentionTextRequest:
            type: object
            properties:
                roomId:
                    type: string
                roomTopic:
                    type: string
                mentionIds:
                    type: array
                    items:
                        type: string
                text:
                    type: string
        wechat.v1.SendGroupTextReply:
            type: object
            properties:
                success:
                    type: boolean
                messageId:
                    type: string
        wechat.v1.SendGroupTextRequest:
            type: object
            properties:
                roomId:
                    type: string
                roomTopic:
                    type: string
                text:
                    type: string
...